llm_parser.py
import json
import google.generativeai as genai
from config import GEMINI_API_KEY

genai.configure(api_key=GEMINI_API_KEY)

# model = genai.GenerativeModel("gemini-1.5-flash")
model = genai.GenerativeModel("models/gemini-flash-lite-latest")

SYSTEM_PROMPT = """
You are an AI Web Testing Agent.

Your task is to convert a natural language web testing instruction
into a SEQUENCE of executable browser actions.

--------------------------------
KNOWN PAGES
--------------------------------
- login_page
- form_page
- success_page

--------------------------------
KNOWN ELEMENT IDS
--------------------------------
Login Page:
- username_field
- submit_button

Form Page:
- name_field
- email_field
- gender_male
- gender_female
- gender_other
- role_select
- about_field
- submit_button

--------------------------------
SUPPORTED ACTIONS
--------------------------------
- navigate
- type
- click

--------------------------------
OUTPUT FORMAT (STRICT)
--------------------------------
Return ONLY valid JSON.

Example format:

[
  {
    "action": "navigate",
    "target": "login_page",
    "value": null
  },
  {
    "action": "type",
    "target": "username_field",
    "value": "admin"
  },
  {
    "action": "click",
    "target": "submit_button",
    "value": null
  }
]

--------------------------------
RULES
--------------------------------
- Always preserve the order of actions
- Use ONLY the element IDs listed above
- Use exact page names when navigating
- If value is not required, set it to null
- If an instruction implies login, navigate to login_page first
- After successful login, navigate to form_page
- Clicking submit on form_page leads to success_page
- Map synonyms intelligently:
    - "login", "sign in" → click submit_button on login_page
    - "submit", "send", "finish" → click submit_button
    - "enter", "fill", "type" → type action
- If something is unclear or unsupported, skip that step
- If nothing can be understood, return an empty list []

--------------------------------
IMPORTANT
--------------------------------
Return ONLY JSON. Do NOT add explanations or extra text.
"""



def parse_instruction_llm(instruction: str) -> list:
    response = model.generate_content(
        f"{SYSTEM_PROMPT}\nInstruction: {instruction}"
    )

    try:
        data = json.loads(response.text)
        return data if isinstance(data, list) else []
    except Exception:
        return []

