requirements.txt
flask
langgraph
langchain
google-generativeai
python-dotenv